% Generated by roxygen2 (4.0.2): do not edit by hand
\name{cal_mss_tc_composite_model}
\alias{cal_mss_tc_composite_model}
\title{Create composite MSS TC model}
\usage{
cal_mss_tc_composite_model(dir)
}
\arguments{
\item{dirname}{character. the directory to the calibration folder}
}
\description{
Create composite MSS TC model with diagnostic figures
}

