% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_oli_tc_aggregate_model.r
\name{cal_oli_tc_aggregate_model}
\alias{cal_oli_tc_aggregate_model}
\title{Create an aggregate oli TC model}
\usage{
cal_oli_tc_aggregate_model(dir, overwrite = F)
}
\arguments{
\item{dirname}{character. the directory to the calibration folder}
}
\description{
Create an aggregate oli TC model with diagnostic figures
}

