% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate_and_composite.r
\name{calibrate_and_composite}
\alias{calibrate_and_composite}
\title{Calibrate MSS imagery to TM and make cloud-free composites}
\usage{
calibrate_and_composite(msswrs1dir, msswrs2dir, tmwrs2dir, oliwrs2dir, index,
  outdir, runname, useareafile, doyears = "all", order = "none",
  overlap = "mean", cores = 2, process, overwrite = F, startday, endday,
  yearadj)
}
\arguments{
\item{msswrs1dir}{character. mss wrs1 directory path}

\item{msswrs2dir}{character. mss wrs2 directory path}

\item{tmwrs2dir}{character. tm wrs2 directory path}

\item{index}{character. spectral index to make composites for. options: "tca", "tcb", "tcg", "tcw"}

\item{outdir}{character. path to output directory}

\item{runname}{character. unique name for the composite set}

\item{useareafile}{character. path to usearea file}

\item{doyears}{??? what years to composite}

\item{order}{character. how to order the images options "sensor_and_doy", "doy", and "none"}

\item{overlap}{character. how to deal with overlapping images. options: "mean"}

\item{cores}{numeric. Number of cores to process with options: 1 or 2}

\item{process}{numeric. integer or vector specifying which processes to run: 1=msscal, 2=mixel}
}
\description{
Calibrate MSS imagery to TM and make cloud-free composites
}

