% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_extension.r
\name{change_extension}
\alias{change_extension}
\title{Changes a file's extension}
\usage{
change_extension(old, new, file)
}
\arguments{
\item{old}{character. old extension}

\item{new}{character. new extension}

\item{file}{character. full path of the file to change extension}
}
\description{
Changes a file's extension
}

