% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_files.r
\name{delete_files}
\alias{delete_files}
\title{Delete files}
\usage{
delete_files(file, reason)
}
\arguments{
\item{file}{character. any file associated with a particular image ID}

\item{reason}{numeric or character. a reason for deleting the image.}
}
\description{
Delete all files associated with a particular image ID
}

