% Generated by roxygen2 (4.0.2): do not edit by hand
\name{doinparallel}
\alias{doinparallel}
\title{Run prepareMSS in parallel}
\usage{
doinparallel(scenedir, demfile, process = seq(1:5), cores = 2)
}
\arguments{
\item{scenedir}{character. scene file path}

\item{demfile}{character. full path to scene-corresponding DEM file}

\item{process}{numeric. integer or vector specifying which processes to run 1=mssunpackr, 2=msswarp, 3=mssdn2rad, 4=mssatcor, 5=msscvm, 6=tmunpackr}

\item{cores}{numeric. number of cores to use for parallel processing}
}
\description{
Run prepareMSS in parallel using
}

