% Generated by roxygen2 (4.0.2): do not edit by hand
\name{eudist}
\alias{eudist}
\title{Earth-Sun distance by day of year}
\usage{
eudist(doy)
}
\arguments{
\item{doy}{numeric. image day-of-year}
}
\description{
Earth-Sun distance by day of year
}
\references{
http://landsathandbook.gsfc.nasa.gov/excel_docs/d.xls
}

