% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_check.r
\name{file_check}
\alias{file_check}
\title{Handles file existence checking and overwriting}
\usage{
file_check(file, output, overwrite)
}
\arguments{
\item{file}{Filename. of file being worked on}

\item{output}{Filename. what output file is it checking for? ("archv.tif", "reflectance.tif", etc)}

\item{overwrite}{logical. if the output file exists should it be deleted}
}
\description{
Handles file existence checking and overwriting
}

