% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_intersecting_images.r
\name{find_intersecting_images}
\alias{find_intersecting_images}
\title{Find intersecting images}
\usage{
find_intersecting_images(files, refimg)
}
\arguments{
\item{files}{character. Array of image filenames}

\item{refimg}{character. filename of selected image that defines spatial extent of intersection search}
}
\description{
From a list of images, find those that interest a given image
}

