% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get_intersection}
\alias{get_intersection}
\title{Find the spatial intersection of a collection of images}
\usage{
get_intersection(files)
}
\arguments{
\item{files}{array of image files names}
}
\description{
Find the spatial intersection of a collection of images
}

