% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_metadata.r
\name{get_metadata}
\alias{get_metadata}
\title{Retrieve Landsat image metadata}
\usage{
get_metadata(file)
}
\arguments{
\item{file}{LPGS-processed Landsat image filename (full system path to file)}
}
\value{
Dataframe with pertinent image information
}
\description{
Uses the image file name to find the corresponding *MTL.txt image metadata file provided with LPSG Landsat images and returns a dataframe with pertinent image information
}

