% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llr_time_machine.r
\name{llr_time_machine}
\alias{llr_time_machine}
\title{Decompress, stack, and reproject LPSG MSS images}
\usage{
llr_time_machine(imgdir, outdir, coordfile)
}
\arguments{
\item{imgdir}{direcory path. full path to directory containing LandsatLinkr annual composites}

\item{outdir}{direcory path. full path to the directory where you want LLR-TimeMachine data to be written}

\item{coordfile}{csv file path. full path to a comma delimited file containing the plot number, x, and y coordinates pixels you want to view in LLR-TimeMachine}
}
\description{
Decompresses, stacks, and optionally reprojects LPGS MSS images recieved from USGS EROS as .tar.gz files
}

