% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ls_profile.r
\name{ls_profile}
\alias{ls_profile}
\title{plot temporal spectral profile for points}
\usage{
ls_profile(msswrs1dir, msswrs2dir, tmwrs2dir, index = "tca", coords = NULL,
  n_random = 10, output = NULL, mode = "evaluate_offset")
}
\arguments{
\item{msswrs1dir}{character. mss wrs1 directory path}

\item{msswrs2dir}{character. mss wrs2 directory path}

\item{tmwrs2dir}{character. tm wrs2 directory path}

\item{index}{character. spectral index to plot temporal profile}

\item{coords}{matrix or data.frame. matrix or dataframe with two row and n columns, column one id x coord, column two is y coord (optional)}

\item{n_random}{numeric. if coords are not given how many random points to plot temporal spectral profile}

\item{output}{character. how to display the profile(s). options "viewer" or "pdf"}

\item{mode.}{character. what to plot. options "apply" or "evaluate"}
}
\description{
plot temporal spectral profile for points
}

