% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ls_segit.r
\name{ls_segit}
\alias{ls_segit}
\title{Piece-wise regression analysis}
\usage{
ls_segit(obj, seg.Z, psi = stop("provide psi"), control = seg.control(),
  model = TRUE, ...)
}
\arguments{
\item{obj}{?look up segmented}

\item{psi}{?look up segmented}

\item{control}{?look up segmented}

\item{model}{?look up segmented}

\item{seg.z}{?look up segmented}
}
\description{
Piece-wise regression analysis
}

