% Generated by roxygen2 (4.0.2): do not edit by hand
\name{make_tc_planes_comparison}
\alias{make_tc_planes_comparison}
\title{Create TC plane comparison plots}
\usage{
make_tc_planes_comparison(tcbsamp, tcgsamp, tcwsamp, outfile)
}
\arguments{
\item{tcb_samp_file}{character. full path name to TC brightness sample}

\item{tcg_samp_file}{character. full path name to TC greenness sample}

\item{tcw_samp_file}{character. full path name to TC wetness sample}

\item{model}{character. which model to use options: "single" or "composite"}

\item{outfile.}{character. define the full path of the .png file to be created}
}
\description{
Create TC plane comparison plots between modeled MSS TC and corresponding TM TC sample
}

