% Generated by roxygen2 (4.0.2): do not edit by hand
\name{make_usearea_file}
\alias{make_usearea_file}
\title{Make a usearea file}
\usage{
make_usearea_file(dir, outdir, xmx, xmn, ymx, ymn)
}
\arguments{
\item{dir}{character. full path name to scene directory example: "E:/mss/wrs2/038029"}

\item{xmx}{numeric.max x coordinate}

\item{xmn}{numeric.min x coordinate}

\item{ymx}{numeric.max y coordinate}

\item{ymn}{numeric.min y coordinate}

\item{outfile}{charcter. full path name for output usearea file}
}
\description{
Make a usearea file for image compositing
}

