% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_usearea_file_bsq.r
\name{make_usearea_file_bsq}
\alias{make_usearea_file_bsq}
\title{Make a usearea file in ENVI (bsq) format}
\usage{
make_usearea_file_bsq(infile, projref)
}
\arguments{
\item{infile}{character. full path name to a usearea mask file}

\item{projref}{charcter. full path to an image file from the scene relevent to the usearea file - can be any .tif file in the "images" directory}
}
\description{
Make a usearea file in ENVI (bsq) format
}

