% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_to_raster.r
\name{matrix_to_raster}
\alias{matrix_to_raster}
\title{Converts a matrix to a raster file}
\usage{
matrix_to_raster(rfile, rmatrix)
}
\arguments{
\item{rfile}{character. full path name of a reference raster file}

\item{rmatrix}{matrix. a numeric 2-d matrix to be converted to a raster}
}
\description{
Converts a matrix to a raster file
}

