% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosaic_dems.r
\name{mosaic_dems}
\alias{mosaic_dems}
\title{Create a DEM mosaic from a direcory of DEM's}
\usage{
mosaic_dems(dir, proj)
}
\arguments{
\item{dir}{character. The path to a directory containing DEM files to be mosaicked}
}
\description{
Create a DEM mosaic from a direcory of DEM's
}

