% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mss_resample.r
\name{mss_resample}
\alias{mss_resample}
\title{Resamples MSS dos_sr and cloudmask images to 30m}
\usage{
mss_resample(file, overwrite = F)
}
\arguments{
\item{file}{character. full path to either an MSS *reflectance.tif or *cloudmask.tif file}
}
\description{
Resamples MSS dos_sr and cloudmask images to 30m
}

