% Generated by roxygen2 (4.0.2): do not edit by hand
\name{mssatcor}
\alias{mssatcor}
\title{Apply dark object subtraction atmospheric correction}
\usage{
mssatcor(file, outtype)
}
\arguments{
\item{file}{The full path name of the *archv file}

\item{outtype}{coded integer designating what units are desired in the output image: 1=DN, 2=radiance, 3=surface reflectance}
}
\description{
Apply dark object subtraction atmospheric correction with option of out putting the image files in units of: DN, radiance, or surface reflectance
}

