% Generated by roxygen2 (4.0.2): do not edit by hand
\name{msscal}
\alias{msscal}
\title{Calibrate MSS images to TM images}
\usage{
msscal(mss_file, tm_file)
}
\arguments{
\item{msswrs2dir}{character. MSS WRS-2 scene directory path}

\item{tmwrs2dir}{character. TM WRS-2 scene directory path}
}
\description{
Calibrate MSS images to TM images using linear regression
}

