% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msscal_single.r
\name{msscal_single}
\alias{msscal_single}
\title{Calibrate MSS images to TM images}
\usage{
msscal_single(mss_file, tm_file)
}
\arguments{
\item{msswrs2dir}{character. MSS WRS-2 scene directory path}

\item{tmwrs2dir}{character. TM WRS-2 scene directory path}
}
\description{
Calibrate MSS images to TM images using linear regression
}

