% Generated by roxygen2 (4.0.2): do not edit by hand
\name{msscost}
\alias{msscost}
\title{Convert DN values to surface reflectance}
\usage{
msscost(file, drkobjv)
}
\arguments{
\item{file}{The full path name of the *archv file}

\item{drkobjv}{dark object values from mssdofinder}
}
\description{
Convert DN values to surface reflectance using the COST model with dark object subtraction
}

