% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/msscvm.r
\name{msscvm}
\alias{msscvm}
\title{Create a cloud and cloud shadow mask for Landsat MSS imagery}
\usage{
msscvm(file, demfile)
}
\arguments{
\item{file}{MSS image filename (full system path to MSS file)}

\item{demfile}{DEM filename (full system path to spatially coincident DEM file)}
}
\value{
A binary raster with the same dimensions as the MSS image where pixels with value 1 represent clear pixel and 0 as obsured by either cloud or cloud shadow
}
\description{
Takes in any numeric value and squares it.
}

