% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/msscvm2.r
\name{msscvm2}
\alias{msscvm2}
\title{Create a cloud and cloud shadow mask for Landsat MSS imagery}
\usage{
msscvm2(file, demfile, test = F)
}
\arguments{
\item{file}{character. MSS reflectance image filename (full system path to MSS file)}

\item{demfile}{character. DEM filename (full system path to spatially coincident DEM file)}
}
\value{
A binary raster with the same dimensions as the MSS image where pixels with value 1 represent clear pixel and 0 as obsured by either cloud or cloud shadow
}
\description{
Takes in any numeric value and squares it.
}

