% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mssdn2rad.r
\name{mssdn2rad}
\alias{mssdn2rad}
\title{Convert MSS DN values to radiance}
\usage{
mssdn2rad(file)
}
\arguments{
\item{file}{character. full file path of *archv.bsq file}
}
\description{
Convert MSS DN values to radiance
}

