% Generated by roxygen2 (4.0.2): do not edit by hand
\name{mssdn2tc}
\alias{mssdn2tc}
\title{Transform original MSS bands to Tasseled Cap}
\usage{
mssdn2tc(file, tcset)
}
\arguments{
\item{file}{The full path name of the *archv file}

\item{Parameter}{indicating which outputs to provide: !list options!}
}
\description{
Transform original MSS bands to Tasseled Cap brightness, greenness, and angle (angle between brightness and greenness). TC yellowness is currently excluded.
}
\details{
Transformation based on coefficents and equation presented in: Pflugmacher, D., Cohen, W. B., & E. Kennedy, R. (2012). Using Landsat-derived disturbance history (1972-2010) to predict current forest structure. Remote Sensing of Environment, 122, 146-165. doi:10.1016/j.rse.2011.09.025
}

