% Generated by roxygen2 (4.0.2): do not edit by hand
\name{mssdofindr}
\alias{mssdofindr}
\title{Determine the Dark Object Values}
\usage{
mssdofindr(file, pngout)
}
\arguments{
\item{file}{The full path name of the *archv file}

\item{pngout}{Logical. If TRUE then a 4-panale .png image file will be created that for each MSS band displays a truncated histogram and the estimated Dark Object Value}
}
\description{
Determine the Dark Object Values in MSS imagery per band
}

