% Generated by roxygen2 (4.0.2): do not edit by hand
\name{mssdos}
\alias{mssdos}
\title{Determine MSS band Dark Object Values}
\usage{
mssdos(file, drkobjv, outtype)
}
\arguments{
\item{file}{The full path name of the *archv file}

\item{pngout}{Logical. If TRUE then a 4-panel .png image file will be created that for each MSS band displays a truncated histogram and the estimated Dark Object Value}
}
\description{
Determine the Dark Object Values in MSS imagery per band
}

