% Generated by roxygen2 (4.0.2): do not edit by hand
\name{mssradcal}
\alias{mssradcal}
\title{Radiometrically calibrate MSS TC to TM TC}
\usage{
mssradcal(mss_rad_file, ref_tc_file)
}
\arguments{
\item{mss_rad_file}{filename full path to MSS radiance file}

\item{ref_tc_file}{filename full path to TM TC reference file}
}
\description{
Spectrally calibrate MSS TC to TM TC
}

