% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msssr2tc.r
\name{msssr2tc}
\alias{msssr2tc}
\title{Create MSS TC from surface reflectance and modeled coefficients}
\usage{
msssr2tc(mss_file, bcoef, gcoef, wcoef, mode)
}
\arguments{
\item{mss_file}{character. full path name to MSS surface reflectance file}

\item{bcoef}{numeric. numeric array containing the tcb coefficients}

\item{gcoef}{numeric. numeric array containing the tcg coefficients}

\item{wcoef}{numeric. numeric array containing the tcw coefficients}

\item{mode.}{character. how to deal with the outputs options: "calibrate" or "apply"}
}
\description{
Create MSS TC from surface reflectance and modeled coefficients
}

