% Generated by roxygen2 (4.0.2): do not edit by hand
\name{mssunpackr}
\alias{mssunpackr}
\title{Decompress, stack, and reproject LPSG MSS images}
\usage{
mssunpackr(file, proj = "default", reso = 60)
}
\arguments{
\item{file}{Filename of LPGS Landsat MSS image filename (full system path to file)}

\item{proj}{CRS projection definition. By default no projection will take place. Optionally specify a CRS projection string or "albers" for the USGS version of Albers Equal Area Conic}

\item{reso}{numeric. the target pixel size for the output image}
}
\description{
Decompresses, stacks, and optionally reprojects LPGS MSS images recieved from USGS EROS as .tar.gz files
}

