% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neatline.r
\name{neatline}
\alias{neatline}
\title{Segment spectral profiles into lines of best fit}
\usage{
neatline(info, tca, tcb, tcg, tcw, outdir, bname)
}
\arguments{
\item{info}{data.frame. chunk information}

\item{tca}{raster. tasseled cap angle raster}

\item{tcb}{raster. tasseled cap brightness raster}

\item{tcg}{raster. tasseled cap greenness raster}

\item{tcw}{raster. tasseled cap wetness raster}

\item{outdir}{character. full path to desired output directory}
}
\description{
Segment spectral profiles into lines of best fit
}

