% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/olical.r
\name{olical}
\alias{olical}
\title{Calibrate OLI images to TM images}
\usage{
olical(oliwrs2dir, tmwrs2dir, cores = 2)
}
\arguments{
\item{oliwrs2dir}{character. oli WRS-2 scene directory path}

\item{tmwrs2dir}{character. TM WRS-2 scene directory path}

\item{cores}{numeric. Number of cores to process with options: 1 or 2}
}
\description{
Calibrate OLI images to TM images using linear regression
}

