% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/olical_single.r
\name{olical_single}
\alias{olical_single}
\title{Calibrate oli images to TM images}
\usage{
olical_single(oli_file, tm_file, overwrite = F)
}
\arguments{
\item{oliwrs2dir}{character. oli WRS-2 scene directory path}

\item{tmwrs2dir}{character. TM WRS-2 scene directory path}
}
\description{
Calibrate oli images to TM images using linear regression
}

