% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/olisr2tc.r
\name{olisr2tc}
\alias{olisr2tc}
\title{Create oli TC from surface reflectance and modeled coefficients}
\usage{
olisr2tc(oli_file, bcoef, gcoef, wcoef, mode, overwrite = F)
}
\arguments{
\item{oli_file}{character. full path name to oli surface reflectance file}

\item{bcoef}{numeric. numeric array containing the tcb coefficients}

\item{gcoef}{numeric. numeric array containing the tcg coefficients}

\item{wcoef}{numeric. numeric array containing the tcw coefficients}

\item{mode.}{character. how to deal with the outputs options: "calibrate" or "apply"}
}
\description{
Create oli TC from surface reflectance and modeled coefficients
}

