% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oliunpackr.r
\name{oliunpackr}
\alias{oliunpackr}
\title{Decompress, stack, and reproject OLI SR images}
\usage{
oliunpackr(file, proj = "default", overwrite = F)
}
\arguments{
\item{file}{character. full path name of the surface reflectance file}

\item{proj}{character. PROJ.4 projection definition.}

\item{overwrite}{logical. True will overwrite the file if it already exists, False will skip processing if output file exists.}
}
\description{
Decompress, stack, and reproject OLI SR images
}

