% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_it.r
\name{plot_it}
\alias{plot_it}
\title{scatterplot of MSS and TM TC transformations}
\usage{
plot_it(samp, coef, r, mss_file, ref_file, index)
}
\arguments{
\item{samp}{data.frame dataframe containing sample of pixel values for MSS and TM TC transformations}

\item{coef}{array linear regression coefficents for linear model between MSS and TM TC transformations}

\item{r}{correlation coefficent for linear relationship between MSS and TM TC transformations}

\item{mss_file}{filename full path to MSS file}

\item{ref_file}{filename full path to TC reference file}

\item{index}{character what index to plot}
}
\description{
create scatterplot of calibrated MSS and TM TC transformations
}

