% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/predict_mss_brt_grn_tca.r
\name{predict_mss_brt_grn_tca}
\alias{predict_mss_brt_grn_tca}
\title{calibrate MSS TC brightness, greenness, and angle}
\usage{
predict_mss_brt_grn_tca(mss_img, ref_img, goods, bins, n)
}
\arguments{
\item{mss_img}{raster* mss radiance brick or stack image containing bands 1-4}

\item{ref_img}{raster* TC wetness reference image}

\item{goods}{array index of pixels to draw a sample from}

\item{bins}{integer number of histogram bins from which to sample}

\item{n}{integer number of pixels to sample per histogram bin}
}
\description{
calibrate MSS TC brightness, greenness, and angle by linear regression between MSS TC components and corresponding reference TC components
}

