% Generated by roxygen2 (4.0.2): do not edit by hand
\name{predict_mss_index}
\alias{predict_mss_index}
\title{Predict MSS TC index}
\usage{
predict_mss_index(refsamp, b1samp, b2samp, b3samp, b4samp, mss_sr_file,
  ref_file, index, outsampfile, samplen)
}
\arguments{
\item{file}{The full path name of the *archv file}

\item{pngout}{Logical. If TRUE then a 4-panel .png image file will be created that for each MSS band displays a truncated histogram and the estimated Dark Object Value}
}
\description{
Predict MSS TC index
}

