% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/predict_mss_wetness.r
\name{predict_mss_wetness}
\alias{predict_mss_wetness}
\title{predict mss TC wetness}
\usage{
predict_mss_wetness(mss_rad_img, ref_tc_img, goods, bins, n)
}
\arguments{
\item{mss_rad_img}{raster* mss radiance brick or stack image containing bands 1-4}

\item{ref_tc_img}{raster* TC wetness reference image}

\item{goods}{array index of pixels to draw a sample from}

\item{bins}{integer number of histogram bins from which to sample}

\item{n}{integer number of pixels to sample per histogram bin}
}
\description{
predicts tasseled cap wetness for mss by linear regression between mss bands 1-4 and TM wetness
}

