% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/predict_oli_index.r
\name{predict_oli_index}
\alias{predict_oli_index}
\title{Predict oli TC index}
\usage{
predict_oli_index(tbl, outsampfile)
}
\arguments{
\item{file}{The full path name of the *archv file}

\item{pngout}{Logical. If TRUE then a 4-panel .png image file will be created that for each oli band displays a truncated histogram and the estimated Dark Object Value}
}
\description{
Predict oli TC index
}

