% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_images.r
\name{prepare_images}
\alias{prepare_images}
\title{Prepare image MSS and TM images for calibration/compositing}
\usage{
prepare_images(scenedir, demfile = NULL, proj = "default",
  process = seq(1:5), cores = 2, overwrite = F)
}
\arguments{
\item{scenedir}{character. scene file path}

\item{demfile}{character. full path to scene-corresponding DEM file}

\item{process}{numeric. integer or vector specifying which processes to run 1=mssunpackr, 2=msswarp, 3=mssdn2rad, 4=mssatcor, 5=msscvm, 6=tmunpackr}

\item{cores}{numeric. number of cores to use for parallel processing}
}
\description{
Prepare image MSS and TM images for calibration/compositing
}

