% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_topo.r
\name{prepare_topo}
\alias{prepare_topo}
\title{Prepare topographic layers: elevation, slope, aspect for use in MSS cloud masking}
\usage{
prepare_topo(imgdir, demfile)
}
\arguments{
\item{imgdir}{character. full path to "images" directory for scene}

\item{demfile}{character. full path to scene-corresponding DEM file}
}
\description{
Prepare topographic layers: elevation, slope, aspect for use in MSS cloud masking
}

