% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/run_calibrate_and_composite.r
\name{run_calibrate_and_composite}
\alias{run_calibrate_and_composite}
\title{GUI for calibrate_and_composite function}
\usage{
run_calibrate_and_composite()
}
\description{
GUI for calibrate_and_composite function
}

