% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sample_it.r
\name{sample_it}
\alias{sample_it}
\title{draw a random sample of image pixels}
\usage{
sample_it(img, bins, n)
}
\arguments{
\item{img}{matrix image matrix to be sampled}

\item{bins}{integer number of histogram bins from which to sample}

\item{n}{integer number of pixels to sample per histogram bin}
}
\description{
draw a histogram equalized random sample of image pixels
}

