% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_projection.r
\name{set_projection}
\alias{set_projection}
\title{Set the projection of raster}
\usage{
set_projection(file)
}
\arguments{
\item{file}{Filename of MSS image with DN values}
}
\description{
Set the projection of a raster from the corresponding *proj.txt file. This solves a problem with incorrect projection parameters for albers when read by 'raster'
}

