% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim_na_rowcol.r
\name{trim_na_rowcol}
\alias{trim_na_rowcol}
\title{Trim NA rows and columns from an image}
\usage{
trim_na_rowcol(imgfile, outimg, maskfile, outmask)
}
\arguments{
\item{imgfile}{character. input image filename}

\item{outimg}{character. output image filename}

\item{maskfile}{character. input mask filename (optional)}

\item{outmask}{character. input mask filename (optional)}
}
\description{
Trim NA rows and columns from an image
}

