% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/write_tc.r
\name{write_tc}
\alias{write_tc}
\title{write out calibrated mss tc files}
\usage{
write_tc(mss_file, index, coef, outfile)
}
\arguments{
\item{mss_file}{filename mss radiance file}

\item{index}{character what tc transformation to be written out}

\item{coef}{array calibrartion coefficents for transformation to tc index}

\item{outfile}{filename full path of the outfile}
}
\description{
write out calibrated mss tc files
}

