% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_mss_tc_aggregate_model.r
\name{cal_mss_tc_aggregate_model}
\alias{cal_mss_tc_aggregate_model}
\title{Create an aggregate MSS TC model}
\usage{
cal_mss_tc_aggregate_model(dir)
}
\arguments{
\item{dirname}{character. the directory to the calibration folder}
}
\description{
Create an aggregate MSS TC model with diagnostic figures
}

