% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_usearea_file.r
\name{make_usearea_file}
\alias{make_usearea_file}
\title{Make a usearea file}
\usage{
make_usearea_file(dir, outfile, xmx, xmn, ymx, ymn)
}
\arguments{
\item{dir}{character. full path name to scene directory example: "E:/mss/wrs2/038029"}

\item{outfile}{charcter. full path of output file}

\item{xmx}{numeric.max x coordinate}

\item{xmn}{numeric.min x coordinate}

\item{ymx}{numeric.max y coordinate}

\item{ymn}{numeric.min y coordinate}
}
\description{
Make a usearea file for image compositing
}

