% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixel.r
\name{mixel}
\alias{mixel}
\title{Composite images}
\usage{
mixel(msswrs1dir, msswrs2dir, tmwrs2dir, oliwrs2dir, index, outdir, runname,
  useareafile, doyears = "all", order = "none", overlap = "mean",
  startday, endday, yearadj = 0)
}
\arguments{
\item{msswrs1dir}{character. list of mss wrs1 directory paths}

\item{msswrs2dir}{character. list of mss wrs2 directory paths}

\item{tmwrs2dir}{character. list of tm wrs2 directory path}

\item{index}{character. spectral index to make composites for. options: "tca", "tcb", "tcg", "tcw"}

\item{outdir}{character. path to output directory}

\item{runname}{character. unique name for the composite set}

\item{useareafile}{character. path to usearea file}

\item{doyears}{??? what years to composite}

\item{order}{character. how to order the images options "sensor_and_doy" and "doy"}

\item{overlap}{character. how to deal with overlapping images. options: "mean"}
}
\description{
Composite images
}

