% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msscal.r
\name{msscal}
\alias{msscal}
\title{Calibrate MSS images to TM images}
\usage{
msscal(msswrs1dir, msswrs2dir, tmwrs2dir, cores = 2)
}
\arguments{
\item{msswrs1dir}{character. MSS WRS-1 scene directory path}

\item{msswrs2dir}{character. MSS WRS-2 scene directory path}

\item{tmwrs2dir}{character. TM WRS-2 scene directory path}

\item{cores}{numeric. Number of cores to process with options: 1 or 2}
}
\description{
Calibrate MSS images to TM images using linear regression
}

