% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msscost.r
\name{msscost}
\alias{msscost}
\title{Convert DN values to surface reflectance}
\usage{
msscost(file, overwrite = F)
}
\arguments{
\item{file}{The full path name of the *archv file}
}
\description{
Convert DN values to surface reflectance using the COST model with dark object subtraction
}

