% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mssdn2refl.r
\name{mssdn2refl}
\alias{mssdn2refl}
\title{Convert DN values to toa reflectance}
\usage{
mssdn2refl(file, overwrite = F)
}
\arguments{
\item{file}{The full path name of the *archv file}
}
\description{
Convert DN values to toa reflectance
}

