% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plot_mss_tm_tc.r
\name{plot_mss_tm_tc}
\alias{plot_mss_tm_tc}
\title{Plot the three TC planes for MSS and TM TC}
\usage{
plot_mss_tm_tc(mss_tc_file, tm_tc_file, bins, n, outfile)
}
\arguments{
\item{mss_tc_file}{character. full filename path to MSS tc file}

\item{tm_tc_file}{character. full filename path to TM TC file}

\item{bins}{integer.}

\item{n}{integer.}

\item{outfile}{}
}
\description{
Plot the three TC planes for MSS and TM TC. It should only be done for coincident images.
}

