% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plot_tc_planes.r
\name{plot_tc_planes}
\alias{plot_tc_planes}
\title{Create TC plane plots}
\usage{
plot_tc_planes(df, plane, sensor)
}
\arguments{
\item{df}{data.frame. data.frame with column names "x" and "y" to be displayed as a 2d scatterplot}

\item{plane}{character. the TC plane to be plotted options: "bg", "bw", "wg"}

\item{sensor}{character. which sensor is this data from options: "mss", "tm"}
}
\description{
Create TC plane plots
}

