% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/run_neatline.r
\name{run_neatline}
\alias{run_neatline}
\title{Run the neatline program}
\usage{
run_neatline(dir, cores)
}
\arguments{
\item{dir}{character. full path to a composite directory}

\item{cores}{numeric. the number of cores to be used for parellel processing}
}
\description{
Run the neatline program
}

