% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tmunpackr.r
\name{tmunpackr}
\alias{tmunpackr}
\title{Decompress, stack, and reproject LPSG MSS images}
\usage{
tmunpackr(file, proj = "default", reso = 30)
}
\arguments{
\item{file}{character. full path name of the surface reflectance file}

\item{outtype}{coded integer designating what units are desired in the output image: 1=DN, 2=radiance, 3=surface reflectance}
}
\description{
Decompresses, stacks, and optionally reprojects LPGS MSS images recieved from USGS EROS as .tar.gz files
}

